refactor: moved types to shared folder
